+++ docs/changelog.rst
@@ -17,6 +17,8 @@ Chula v0.13.0 (dev)
 
 *Still under development*
 
+- Add support for using the Eventlet wsgi provider.
+
 :Source: http://github.com/jmcfarlane/chula
 
 Chula v0.12.0 (latest)